feat: change structure of code
